Move BoardSquare into Piece.js
